requirements.txt
# Invoice processing & parsing
pandas>=2.0.0

# Jupyter kernel for .venv
ipykernel>=6.0.0

# PDF invoice parsing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# LangGraph pipeline
langgraph>=0.2.0

# LLM extraction (TXT/PDF) + env
python-dotenv>=1.0.0
openai>=1.0.0

# Tests
pytest>=7.0.0

# Optional: Streamlit UI
streamlit>=1.28.0
